# openapi.yaml
openapi: 3.1.0
info:
  title: Badminton Court Booking API
  # The version of this DOCUMENT. Not the same as the v1 in the URL.
  version: 1.2.1
  description: |
    Books badminton courts at a university sports hall.

    It is written for three kinds of caller: **students**, who browse the
    courts and book, read and cancel their own slots; **administrators**,
    who manage the courts themselves and can act on any booking; and a
    **scheduled job** that runs nightly with no user present to tidy up
    expired bookings.

    ## Conventions

    All timestamps are RFC 3339 with an explicit UTC offset, for example
    `2026-03-04T12:10:04+07:00`. A booking covers a half-open interval: a
    booking that ends at 19:00 does not overlap one that starts at 19:00.

    All identifiers are opaque and server-generated, prefixed by the kind of
    thing they name — `crt_` for a court, `bkg_` for a booking. They are
    stable forever. Clients MUST NOT parse them or infer ordering from them.

    Clients MUST ignore unknown response fields. Clients MUST treat an
    unrecognised enumeration value as the nearest non-terminal state and
    MUST NOT crash on it. These two sentences are what make every additive
    change to this interface legal.

    This service carries no monetary values. Courts are booked, not bought.

    ## Access control

    Every operation except `GET /health` requires an OAuth 2 access token,
...

...
       declares — a missing scope is `403`;
...

...

...

    ## Idempotency
...
servers:
...
security:
  - oauth2: [courts:read]
paths:
...
  /bookings:
    # -----------------------------------------------------------------
    # A collection constrained by the caller, not by a query parameter.
    # There is no ownerId filter here on purpose: who the caller is comes
    # from the token, so a client cannot ask for somebody else's page.
    # -----------------------------------------------------------------
    get:
      operationId: listBookings
      summary: List the bookings visible to the caller
      tags: [Bookings]
      security:
        - oauth2: [bookings:read]
      description: |
        Returns the bookings this caller may see, and nothing else. A
        caller holding `bookings:read` alone sees the bookings they
        created. A caller also holding `bookings:fulfil` sees every
        booking, because that scope is granted only to administrators
        and the scheduled job.

        The constraint is applied inside the query, so a page of `limit`
        items is a page of items the caller can actually see.
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
        - name: status
          in: query
          required: false
          description: Filter to bookings in one status.
          schema:
            type: string
            enum: [confirmed, cancelled]
        - name: limit
          in: query
          required: false
          description: Maximum number of items to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          required: false
          description: Opaque cursor from a previous page's nextCursor.
          schema: { type: string }
      responses:
        '304':
          $ref: '#/components/responses/NotModified'
        '200':
          description: A page of bookings visible to the caller.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Booking' }
                  nextCursor:
                    type: string
                    description: Absent on the last page.
                    example: YmtnXzUxRmE5M2NE
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    # -----------------------------------------------------------------
    # The unsafe, consequential operation. Required for L2.
    # The idempotency key is declared REQUIRED, so it is impossible to
    # read this document and not know about it.
    # -----------------------------------------------------------------
    post:
      operationId: createBooking
      summary: Create a badminton court booking
      tags: [Bookings]
      security:
        - oauth2: [bookings:write]
      description: |
        Creates a badminton court booking.

        This is a consequential operation: the same booking intent MUST NOT
        create multiple bookings when the client retries the request.
        Clients MUST generate an Idempotency-Key when the user confirms
        the booking and MUST reuse the same key for retries of that intent.
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            # A version-4 UUID with hyphens, as stated under Idempotency above.
            # The pattern is the whole constraint; format: uuid is not used
            # because it admits every UUID version and says less than this.
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
          example: 0f7c1b9e-3d21-4a6f-9c05-8e2b7d41a9f0
          description: |
            Generated by the client at the moment the user confirms the
            action, and reused unchanged for every retry of that same
            intent. Retained 24 hours. Reuse with a different body
            returns 409 idempotency-key-reuse.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NewBooking' }
      responses:
        '201':
          description: |
            Created — or, on a retry with the same key and the same
            body, the replayed original response.
          headers:
            Location:
              description: URI of the created booking.
              schema: { type: string }
              # Without this example the mock server serves the literal
              # word "string" in the header. Every example you omit is a
              # place where the mock stops being useful to a client
              # developer — which is the whole point of having one.
              example: /v1/bookings/bkg_51Fa93cD
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }

  # -------------------------------------------------------------------
  # The operations that name a single booking. A valid token carrying
  # bookings:read is not enough to read any of them: the handler also
  # checks the relationship between the caller and the booking, and
  # answers 404 when there is none.
  # -------------------------------------------------------------------
  /bookings/{bookingId}:
    parameters:
      - $ref: '#/components/parameters/BookingId'
    get:
      operationId: getBooking
      summary: Fetch one booking
      tags: [Bookings]
      security:
        - oauth2: [bookings:read]
      description: |
        Returns the booking only when the caller created it, or holds
        `bookings:fulfil`. A booking that exists but belongs to another
        caller is answered exactly as a booking that does not exist:
        same status, same `type`, same body. The two cases are
        deliberately indistinguishable so booking identifiers cannot be
        enumerated.
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '304':
          $ref: '#/components/responses/NotModified'
        '200':
          description: The booking.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # -------------------------------------------------------------------
  # Rule 5 again — the cancellation is a noun sub-resource, so it can
  # carry a reason, return who and when, and be read back later.
  # -------------------------------------------------------------------
  /bookings/{bookingId}/cancellation:
    parameters:
      - $ref: '#/components/parameters/BookingId'
    post:
      operationId: cancelBooking
      summary: Cancel a booking
      tags: [Bookings]
      security:
        - oauth2: [bookings:write]
      description: |
        Cancels a booking the caller created. The ownership check runs
        before anything is written, so a refused request leaves the
        booking exactly as it was.

        Re-sending this when the booking is ALREADY cancelled returns
        200 with the existing record, not 409. The requested end state
        already holds and the caller's intent has been satisfied.

        A booking belonging to another caller is answered 404, the same
        as one that does not exist.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 1
                  maxLength: 500
                  # Control characters are excluded, NUL among them. The
                  # constraint is stated here because it is a real one: the
                  # reason is stored as text, and a NUL byte cannot be held
                  # in a Postgres text column. Leaving it unsaid let a
                  # schema-compliant request reach the database and come
                  # back as a 500.
                  pattern: '^[^\u0000-\u001F\u007F]+$'
                  description: |
                    Why the booking is being cancelled. Free text, one to
                    500 characters, no control characters.
                  example: Cannot make the booked slot
      responses:
        '200':
          description: Already cancelled. The existing record is returned.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cancellation' }
        '201':
          description: Cancelled.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cancellation' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
...
  securitySchemes:
    # One scheme carrying both flows, rather than two schemes. The
    # scheduled job and the two web clients present the same kind of token
    # to the same API; what differs is only how each one obtained it. Split
    # across two schemes, the client-credentials half was referenced by no
    # operation, which the linter reported as an unused component — an
    # accurate description of a flow nothing could actually use.
    oauth2:
      type: oauth2
      description: |
        Access tokens are issued by the Authorisation Server. Public
        clients use Authorization Code + PKCE with no client secret. The
        scheduled cleanup job has no user present, so it authenticates
        itself with Client Credentials and a secret held in a secret
        manager — never in this repository.
      flows:
        authorizationCode:
          authorizationUrl: https://auth.example.com/realms/badminton/protocol/openid-connect/auth
          tokenUrl: https://auth.example.com/realms/badminton/protocol/openid-connect/token
          refreshUrl: https://auth.example.com/realms/badminton/protocol/openid-connect/token
          scopes:
            courts:read: Browse badminton courts and availability
            courts:write: Create, update, or retire badminton courts
            bookings:read: Read bookings visible to the principal
            bookings:write: Create and cancel own bookings
            bookings:fulfil: Confirm, reject, or manage all bookings
        clientCredentials:
          tokenUrl: https://auth.example.com/realms/badminton/protocol/openid-connect/token
          scopes:
            bookings:read: Read bookings for the scheduled cleanup job
            bookings:fulfil: Manage all bookings for the scheduled cleanup job
  parameters:
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: |
        The `ETag` of the version the client already holds. When it still
        matches, the service answers `304 Not Modified` with no body —
        which is a successful read, not a failure.

        A client polling a collection sends this on every cycle. Without
        it the service has no way to tell that the caller already has
        the current version, and every poll transfers the whole list.
      schema:
        type: string
        example: '"v8m6WCTfmEMOvl8xpz5syHmURr7"'
    BookingId:
      name: bookingId
      in: path
      required: true
      description: Opaque, server-generated. Never a database integer.
      schema:
        type: string
        pattern: '^bkg_[A-Za-z0-9]{3,}$'
      example: bkg_51Fa93cD
  schemas:
...
    NewBooking:
      type: object
      required: [courtId, startTime, endTime]
      properties:
        courtId:
          type: string
          pattern: '^crt_[A-Za-z0-9]{3,}$'
          description: ID of the badminton court to book.
          example: crt_51Fa93cD
        startTime:
          type: string
          format: date-time
          description: Start time of the badminton court booking.
          example: '2026-08-29T19:00:00+07:00'
        endTime:
          type: string
          format: date-time
          description: End time of the badminton court booking.
          example: '2026-08-29T20:00:00+07:00'
    Booking:
      type: object
      required: [id, courtId, startTime, endTime, status]
      properties:
        id:
          type: string
          pattern: '^bkg_[A-Za-z0-9]{3,}$'
          description: Opaque, server-generated, stable forever.
          example: bkg_51Fa93cD
        courtId:
          type: string
          description: ID of the booked badminton court.
          example: crt_51Fa93cD
        startTime:
          type: string
          format: date-time
          description: Start time of the booking.
          example: '2026-08-29T19:00:00+07:00'
        endTime:
          type: string
          format: date-time
          description: End time of the booking.
          example: '2026-08-29T20:00:00+07:00'
        status:
          type: string
          enum: [confirmed, cancelled]
          example: confirmed
        bookedBy:
          type: string
          description: |
            The subject identifier of the caller who created this booking.

            Present only for callers holding `bookings:fulfil` —
            administrators and the scheduled cleanup job — who need to know
            whose booking they are acting on. A student reading their own
            booking does not receive this field, because it would tell them
            nothing they do not already know while making every other
            caller's identifier reachable through a shared representation.
          example: 6f1c9d2a-4b77-4c1e-9f0a-2d8b51e3ac10
    Cancellation:
      type: object
      required: [bookingId, reason, cancelledAt]
      properties:
        bookingId:
          type: string
          pattern: '^bkg_[A-Za-z0-9]{3,}$'
          example: bkg_51Fa93cD
        reason:
          type: string
          example: Cannot make the booked slot
        cancelledAt:
          type: string
          format: date-time
          example: '2026-08-28T09:14:00+07:00'
...
    Problem:
      type: object
      required: [type, title, status]
      # RFC 9457 permits EXTENSION MEMBERS: your own fields alongside the
      # five below. They are what turns a merely correct error into an
      # interface a client can build a good experience on, and they are
      # explicitly assessed in A1 — so the schema must allow them.
      # Remove this line and your own extension members become lint
      # errors, which is exactly how a team ends up with useless errors.
      additionalProperties: true
      properties:
        type:
          type: string
          format: uri
          description: |
            Stable URI identifying the KIND of problem. Client code
            branches on this exact string, so it must never change
            once published.
          example: https://api.example.com/problems/illegal-transition
        title:
          type: string
          description: |
            Short, human-readable, FIXED for a given type. Written for
            a developer reading a log, not for an end user.
          example: That status change is not permitted
        status:
          type: integer
          description: The HTTP status, repeated so the document stays
            meaningful when logged away from its response.
          example: 409
        detail:
          type: string
          description: |
            Specific to this occurrence. Never a stack trace, an SQL
            fragment, or an internal hostname.
          example: thg_51Fa93cD is draft; retirement is refused.
        instance:
          type: string
          description: |
            The URI of this specific occurrence. In Chapter 9 the
            correlation identifier goes here.
          example: /v1/things/thg_51Fa93cD/retirement
        invalid-params:
          type: array
          description: |
            Which fields were refused and why — the extension member
            RFC 9457 uses in its own example. Present on schema
            failures (400) and on cross-field rules (422).

            Each entry names one field and carries a reason written for
            the person filling the form in, so a client can place the
            message on that field rather than inventing wording for a
            rule only this service knows. A client that receives this
            member MUST NOT assume the list is exhaustive: fields are
            reported as they are checked.
          items:
            type: object
            required: [name, reason]
            properties:
              name:
                type: string
                description: |
                  The field the problem is about, as named in the
                  request body or query string.
                example: endTime
              reason:
                type: string
                description: Why it was refused, in domain terms.
                example: The end time must be after the start time
  headers:
    ETag:
      description: |
        Opaque version marker for this exact representation. A client
        stores it and sends it back in `If-None-Match` on the next read
        (A.7), and in `If-Match` on a write against the same entity
        (A.8).

        Derived from the representation, so it changes when, and only
        when, what the client would receive changes. It is strong: two
        responses carrying the same tag are byte-for-byte identical.
      required: true
      schema:
        type: string
        example: '"v8m6WCTfmEMOvl8xpz5syHmURr7"'
    CacheControl:
      description: |
        `private, no-cache`. These representations are shaped by who
        asked — a student's booking list is theirs alone — so a shared
        cache must not reuse one for somebody else. `no-cache` means
        "revalidate before reusing", which is the conversation `ETag`
        and `If-None-Match` are for.
      schema:
        type: string
        example: private, no-cache
  responses:
    NotModified:
      description: |
        The representation has not changed since the version named by
        `If-None-Match`. No body is sent.

        This is a successful read. A client treats it as confirmation
        that what it already holds is current — clearing any stale
        marker — and not as a failure.
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
    BadRequest:
      description: The request was malformed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.example.com/problems/malformed-request
            title: The request could not be parsed
            status: 400
            detail: One or more fields do not match the documented schema
            invalid-params:
              - name: startTime
                reason: Give a start date and time
    Unauthorized:
      description: The token is absent, unverifiable, or expired.
      headers:
        WWW-Authenticate:
          schema: { type: string }
          example: 'Bearer error="invalid_token"'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: The token is valid but does not carry the scope this operation requires.
      headers:
        WWW-Authenticate:
          schema: { type: string }
          example: 'Bearer error="insufficient_scope", scope="bookings:write"'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: |
        The object does not exist, or it exists and is not accessible to this caller.
        Both conditions are answered identically so identifiers cannot be enumerated.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.example.com/problems/not-found
            title: Resource not found
            status: 404
    Conflict:
      description: |
        The request was valid but refused by a business rule. This is a
        domain OUTCOME, not a server error. Do not retry it.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.example.com/problems/court-slot-unavailable
            title: The badminton court is already booked for this time slot
            status: 409
    UnprocessableEntity:
      description: Well-formed but semantically invalid.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.example.com/problems/validation-failed
            title: One or more fields are invalid
            status: 422